Proyecto_Final_Web_3/App web Nro 3 - Sistema de entrega/sistema_entrega_back_end/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-fsm==2.8.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
PyJWT==2.4.0
django-cors-headers==3.13.0
